Use long double instead of double

+++ core/src/linalg.cpp
@@ -5,84 +5,84 @@ using namespace Eigen;
 using namespace linalg;
 
 namespace linalg {
-  DVector<double> mean(
-    const DMatrix<double> &data
+  DVector<long double> mean(
+    const DMatrix<long double> &data
     ) {
     return data.colwise().mean();
   }
 
-  DMatrix<double> outer_product(
-    const DVector<double> &a,
-    const DVector<double> &b
+  DMatrix<long double> outer_product(
+    const DVector<long double> &a,
+    const DVector<long double> &b
     ) {
     return a * b.transpose();
   }
 
-  DMatrix<double> outer_square(
-    const DVector<double> &a
+  DMatrix<long double> outer_square(
+    const DVector<long double> &a
     ) {
     return outer_product(a, a);
   }
 
-  double inner_product(
-    const DVector<double> &a,
-    const DVector<double> &b,
-    const DMatrix<double> &weights) {
+  long double inner_product(
+    const DVector<long double> &a,
+    const DVector<long double> &b,
+    const DMatrix<long double> &weights) {
     return a.transpose() * weights * b;
   }
 
-  double inner_product(
-    const DVector<double> &a,
-    const DVector<double> &b
+  long double inner_product(
+    const DVector<long double> &a,
+    const DVector<long double> &b
     ) {
-    return inner_product(a, b, DMatrix<double>::Identity(a.size(), b.size()));
+    return inner_product(a, b, DMatrix<long double>::Identity(a.size(), b.size()));
   }
 
-  double inner_square(
-    const DVector<double> &a,
-    const DMatrix<double> &weights
+  long double inner_square(
+    const DVector<long double> &a,
+    const DMatrix<long double> &weights
     ) {
     return inner_product(a, a, weights);
   }
 
-  DMatrix<double> inner_product(
-    const DMatrix<double> &a,
-    const DMatrix<double> &b,
-    const DMatrix<double> &weights) {
+  DMatrix<long double> inner_product(
+    const DMatrix<long double> &a,
+    const DMatrix<long double> &b,
+    const DMatrix<long double> &weights) {
     return (a.transpose() * weights * b);
   }
 
-  DMatrix<double> inner_product(
-    const DMatrix<double> &a,
-    const DMatrix<double> &b
+  DMatrix<long double> inner_product(
+    const DMatrix<long double> &a,
+    const DMatrix<long double> &b
     ) {
-    return inner_product(a, b, DMatrix<double>::Identity(a.rows(), b.cols()));
+    return inner_product(a, b, DMatrix<long double>::Identity(a.rows(), b.cols()));
   }
 
-  DMatrix<double> inner_square(
-    const DMatrix<double> &m,
-    const DMatrix<double> &weights
+  DMatrix<long double> inner_square(
+    const DMatrix<long double> &m,
+    const DMatrix<long double> &weights
     ) {
     return inner_product(m, m, weights);
   }
 
-  double determinant(
-    const DMatrix<double> &m
+  long double determinant(
+    const DMatrix<long double> &m
     ) {
     return m.determinant();
   }
 
-  DMatrix<double> inverse(
-    const DMatrix<double> &m
+  DMatrix<long double> inverse(
+    const DMatrix<long double> &m
     ) {
-    DMatrix<double> inverse = m.inverse();
+    DMatrix<long double> inverse = m.inverse();
     assert(inverse.allFinite() && "Given matrix is not invertible");
     return inverse;
   }
 
-  std::tuple<DVector<double>, DMatrix<double> > sort_eigen(
-    const DVector<double> &values,
-    const DMatrix<double> &vectors
+  std::tuple<DVector<long double>, DMatrix<long double> > sort_eigen(
+    const DVector<long double> &values,
+    const DMatrix<long double> &vectors
     ) {
     DVector<int> idx = DVector<int>::Zero(values.size());
 
@@ -90,7 +90,7 @@ namespace linalg {
       idx[i] = i;
     }
 
-    std::sort(idx.data(), idx.data() + idx.size(), [&values](double a, double b)
+    std::sort(idx.data(), idx.data() + idx.size(), [&values](long double a, long double b)
     {
       return abs(values.row(a).value()) < abs(values.row(b).value());
     });
@@ -98,32 +98,32 @@ namespace linalg {
     return std::make_tuple(values(idx), vectors(all, idx));
   }
 
-  std::tuple<DVector<double>, DMatrix<double> > eigen(
-    const DMatrix<double> &m
+  std::tuple<DVector<long double>, DMatrix<long double> > eigen(
+    const DMatrix<long double> &m
     ) {
     if (!m.isApprox(m.transpose())) {
       LOG_INFO << "Non-symmetric matrix detected, using general eigenvalue solver" << std::endl;
 
-      EigenSolver<DMatrix<double> > solver(m);
-      DVector<double> values = solver.eigenvalues().real();
-      DMatrix<double> vectors = solver.eigenvectors().real();
+      EigenSolver<DMatrix<long double> > solver(m);
+      DVector<long double> values = solver.eigenvalues().real();
+      DMatrix<long double> vectors = solver.eigenvectors().real();
 
       return sort_eigen(values, vectors);
     }
 
     LOG_INFO << "Symmetric matrix detected, using self-adjoint eigenvalue solver" << std::endl;
 
-    SelfAdjointEigenSolver<DMatrix<double> > solver(m);
-    DVector<double> values = solver.eigenvalues().real();
-    DMatrix<double> vectors = solver.eigenvectors().real();
+    SelfAdjointEigenSolver<DMatrix<long double> > solver(m);
+    DVector<long double> values = solver.eigenvalues().real();
+    DMatrix<long double> vectors = solver.eigenvectors().real();
 
     return std::make_tuple(values, vectors);
   }
 
   bool collinear(
-    const DVector<double> &a,
-    const DVector<double> &b) {
-    double tolerance = 0.0001;
+    const DVector<long double> &a,
+    const DVector<long double> &b) {
+    long double tolerance = 0.0001;
 
     return fabs(inner_product(a, b) / (a.norm() * b.norm()) - 1.0) < tolerance;
   }
